feat: seperate components into 3 files, style header
