+++ src/app/components/header/header.component.html
@@ -1,0 +1,20 @@
+<header class="header-content-wrapper">
+  <div class="logo-container">
+    <a
+      routerLink="/"
+      routerLinkActive="active"
+      ariaCurrentWhenActive="page"
+      aria-label="Logo"
+    >
+      <img class="logo" src="assets/logo.svg" alt="site logo" />
+    </a>
+  </div>
+  @if (isAuthenticated()) {
+    <app-profile-status-bar></app-profile-status-bar>
+  } @else {
+    <div class="buttons-container">
+      <button mat-button (click)="toLogin()">Login</button>
+      <button mat-button (click)="toRegistration()">Register</button>
+    </div>
+  }
+</header>
