+++ src/app/components/header/header.component.scss
@@ -1,0 +1,48 @@
+@import "../../../styles/colors.scss";
+@import "../../../styles/breakpoints";
+
+::ng-deep app-header {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  background-color: $header-yellow;
+}
+
+.header-content-wrapper {
+  width: min(90vw, 1200px);
+  margin-inline: auto;
+  height: 80px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.logo-container {
+  width: 50px;
+}
+
+.logo {
+  width: 50px;
+}
+
+.main-bar {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+}
+
+.buttons-container {
+  display: flex;
+  gap: 20px;
+
+  button {
+    height: 80px;
+
+    @include hover {
+      &:hover {
+        border-bottom: 3px solid black;
+      }
+    }
+  }
+}
