+++ src/app/components/profile-status-bar/profile-status-bar.component.html
@@ -1,0 +1,12 @@
+<div class="login-info-container">
+  <p>Welcome again!</p>
+  <img
+    src="assets/login.svg"
+    alt="icon of logined account"
+    [matMenuTriggerFor]="beforeMenu"
+  />
+  <mat-menu #beforeMenu="matMenu">
+    <!-- <button mat-menu-item (click)="openProfile()">Profile</button> -->
+    <button mat-menu-item (click)="logout()">Logout</button>
+  </mat-menu>
+</div>
