+++ src/app/components/profile-status-bar/profile-status-bar.component.scss
@@ -1,0 +1,22 @@
+.login-info-container {
+  font-weight: 500;
+  display: flex;
+  gap: 30px;
+  align-items: center;
+  justify-content: flex-end;
+
+  span {
+    text-transform: capitalize;
+    font-size: 16px;
+  }
+
+  img {
+    cursor: pointer;
+    height: 40px;
+  }
+
+  p {
+    margin: 0;
+    cursor: default;
+  }
+}
