+++ src/styles.scss
@@ -1,4 +1,12 @@
 /* You can add global styles to this file, and also import other style files */
+@import "styles/_colors";
+@import "styles/breakpoints";
 
-html, body { height: 100%; }
-body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
+html,
+body {
+  height: 100%;
+}
+body {
+  margin: 0;
+  font-family: Roboto, "Helvetica Neue", sans-serif;
+}
