+++ src/styles/_breakpoints.scss
@@ -1,0 +1,23 @@
+$breakpoints: (
+  "mobile": 450px,
+  "tablet": 768px,
+  "desktop": 1440px,
+);
+
+@mixin mobile {
+  @media screen and (max-width: map-get($breakpoints, "mobile")) {
+    @content;
+  }
+}
+
+@mixin tablet {
+  @media screen and (max-width: map-get($breakpoints, "tablet")) {
+    @content;
+  }
+}
+
+@mixin hover {
+  @media (hover: hover) {
+    @content;
+  }
+}
