+++ src/styles/_colors.scss
@@ -1,0 +1,5 @@
+$text: #000;
+$accent-text: #545454;
+$symbols: #d44535;
+$main-yellow: #ffda1f;
+$header-yellow: #efc81e;
